fw-update: Add config json and parser for firmware update
The documentation for the config json is available in
fw-update/examples/fw_update.json. The config json parser parses the
json and provides config information to the PLDM firmware update
code for device inventory, firmware inventory and to populate the
args for the messages in the update message registry.

The fw_update config json is optional and does not affect firmware
update. Firmware update message registries will be created with
component identifier instead of component name is config json is not
provided.

Tested: Unit tests added

Signed-off-by: Tom Joseph <rushtotom@gmail.com>
Change-Id: I64bacd643ed6a501aa31aea98635642d817e4b66