watchdog: Bump Watchdog timeout to 60 sec
The watchdog timeout of 10 sec was arbitrarily chosen. PLDM makes
synchronous D-Bus calls and it was observed that when HMC was hogged,
D-Bus client calls made by PLDM took longer than 10 sec.

The reply_timeout of D-Bus broker is ~25 sec and based on that the
watchdog timeout is bumped to 60 sec. This will ensure that a sync
PLDM client D-Bus call timing out doesn't cause PLDM daemon to
be reset by watchdog.

Signed-off-by: Tom Joseph <rushtotom@gmail.com>