fw-update: Handle staging directory cleanup
pldmd handles the cleanup of the staging directory once the firmware
when the new package is uploaded. The activation D-Bus object is
created for every path when the package is handled. When a new package
is uploaded, the previous package and activation info is cleaned up.

Fixes nvbug https://nvbugs/3672739

Tested:

Steps:
1. Configure targets for HMC_Firmware
2. Upload ec_fw image
3. Upload ec_fw image again
4. Verify /tmp/images is cleaned up after every upload

Signed-off-by: Tom Joseph <rushtotom@gmail.com>