fwupdate: Clear pldm package after update completion
The staging directory for firmware update is 200M and pldmd clears
the staging directory lazily when a new package is uploaded to the
staging directory, so there is a window where the previous package
and new package is stored in the staging directory.

For QS0 the Vulcan bundle will be ~106M and to handle this pldmd
will clear the firmware update package post firmware update.

This patch removes pldm package on update success, failure or when
there is no matching updates.

Unit Tests:
  1. Verify pldm package is removed in success, failure and when there
     is no matching updates.
  2. Perform update back to back and make sure update is success and no
     stale file remains.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-4924

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>