PLDM Stack Robustness: Additional Enhancements part 2
An enhancement designed to safeguard the package against damage in the event of a truncated component.
An attempt to read such a component from the package may lead to an effort to read a set of
bytes beyond the package's boundaries, triggering the state of the package to be set to std::ios::failbit.
This, in turn, could potentially block the ability to read other components from the package.

Unit Tests:
Run firmware update procedure with a firmware package missing 8K data at the end.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-8393

Signed-off-by: Robert Papros <rpapros@nvidia.com>