mctp-demux-daemon: Handle MCTP message tag and tag owner in PLDM repository
The requirements from the MCTP specification for tag owner and message tags are as below. This patch introduces the capability of PLDM to handle the message tag and tag owner. The API between PLDM daemon and mctp-demux-daemon is updated to include the tag owner and the message tag. When BMC/HMC is the PLDM requester the tag owner bit should be set to 1 and message tag bits should be set to 0. When BMC/BMC is the PLD responder the tag owner bit should be set to 0 and message tag bits should be returned as how it was in the request.

```
<Tag owner & Message Tag Byte><EID><Command Type><PLDM message - New format for the Tx/Rx with the mctp-demux-daemon

Tag Owner & Message Tag Byte:
Bit 7:4 - Reserved
Bit 3 - Tag owner bit
Bit 2:0 - Message tag
```

```
For MCTP control messages, the Rq bit shall be set to 1b if the message is a “command” or Request
message and 0b if the message is a Response message. For Datagram and Broadcast messages, the
Rq bit shall always be set to 1b. MCTP Control messages that have unexpected or incorrect flag bit
values shall be silently discarded by the receiver of the message.
```

```
The TO (Tag Owner) bit identifies whether the message tag was originated by the endpoint that is the source of the message or by the endpoint that is the destination of the message. The Message Tag field is generated and tracked independently for each value of the Tag Owner bit. MCTP message types may overlay this bit with additional meaning, for example using it to differentiate between "request" messages and "response" messages. Set to 1b to indicate that the source of the message originated the message tag.
```

```
(Message tag) Field that, along with the Source Endpoint IDs and the Tag Owner (TO) field, identifies a unique message at the MCTP transport level. Whether other elements, such as portions of the MCTP Message Data field, are also used for uniquely identifying instances or tracking retries of a message is dependent on the message type. A source endpoint is allowed to interleave packets from multiple messages to
the same destination endpoint concurrently, provided that each of the messages has a unique message tag. When request/response message exchange is used and the Tag Owner (TO) bit is set to 1 in the request, a responder should return the same Message Tag with the Message Tag Owner bit cleared to 0 in the corresponding response Message. For messages that are split up into multiple packets, the Tag Owner (TO) and
Message Tag bits remain the same for all packets from the SOM through the EOM.
```

Fixes nvbug https://jirasw.nvidia.com/browse/DGXOPENBMC-10718

Signed-off-by: Tom Joseph <rushtotom@gmail.com>