usb vs spi
add usb bindings in the list of bindings so if there are multiple
EID's for the same target the fastest transport would
be used.
Tested on GB200 platform that has target over SPI as EID 0
and USB as EID 15. EID 15 was correctly chosen

Signed-off-by: Tom Joseph <rushtotom@gmail.com>
Signed-off-by: Radivoje (Ogi) Jovanovic <rjovanovic@nvidia.com>
Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-11596