fw_update: Fix picking the fastest path to a PLDM FD
The logic to pick the fastest path is documented in the FW update platform guide.

https://gitlab-master.nvidia.com/dgx/bmc/docs/-/blob/tjoseph-develop-patch-67740/nvidia/fw_update_platform_guide.md?ref_type=heads#pldm-fw-update-path

MCTP over USB support was added for a new platform and code changes
were added to support the new MCTP Binding type.

Ideally the commit[1] would have sufficed (which was copy paste from PLDM T2), but the
namespace was wrong for binding type which caused the fastest path selection
logic to break. Fixed the PLDM T2 code with commit[2].

[1] https://gitlab-master.nvidia.com/dgx/bmc/pldm/-/commit/ce87813c54928f28237bb7610bac47a37ddf6b86
[2] https://gitlab-master.nvidia.com/dgx/bmc/pldm/-/commit/dc9882eda0b9104aea3c50d73106bf87957a0f92

As I was debugging this issue, noticed that the code was really hard to read
and took this opportunity to refactor the code. There is no change in the
logic to pick the fatest path. The journal logs will clearly indicate which
EID was picked when two endpoints are detected for the same PLDM device.

Tested:

1. Restart mctp-usb-ctrl and verify descriptor erase for endpoint via USB is skipped
2. Verify firmware update picks correct EID after restarting mctp-usb-ctrl
3. Verify firmware update picks correct EID after restarting mctp-spi-ctrl
4. Verify firmware update picks correct EID after restarting pldmd
5. Verify firmware update switches to fatest EID after discovering slower path first.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-11596

Signed-off-by: Tom Joseph <rushtotom@gmail.com>