fw-update: Marking component as failed on unresponsive CancelUpdateComponent request
Currently, the HMC stalls whenever a component does not respond to CancelUpdateComponent request leading to a timeout instead of failing early.
HMC wasn't setting UpdateFailed on the component previously when this scenario occurred which caused the stall.

This patch sets UpdateFailed on the component whenever a component is unresponsive to the CancelUpdateComponent request

Fixes Jira https://

Signed-off-by: P Arun Kumar Reddy parunkumarr@nvidia.com