fw_update: Fix property matches in config JSON
When matching properties in a given interface, we need to ensure that
all properties in the match key of the config JSON are considered for a
match, not just any one. A list of properties under a given interface
act as an AND of all the properties specified.

Fixes NVBUG https://nvbugspro.nvidia.com/bug/4840111

Tested:
Tested that UUID based, EID based and I2C address/bus based matches work
fine on a GB200NVL-BMC.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>