Add null check before stopping and resetting reqFwDataTimer to prevent potential crashes
In cases where PLDM UA receives a TransferComplete message without first receiving a RequestUpdate,
PLDMD was crashing because reqFwDataTimer was not initialized, and it attempted to stop the uninitialized timer.
This commit adds a null check to ensure the timer is only stopped and reset if it has initialized.

Tests:
1. Verified Vbios FW Update StageAndActivate, when an image is already staged in the inactive slot.
2. Verified Vbios FW Update StageOnly, when the same image is already staged in the inactive slot.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4850200

Signed-off-by: Rajeev Ranjan's avatarRajeev Ranjan <rajeeranjan@nvidia.com>