platform-mc: set default value to shm in numeric sensor ctor
The update function of non-priority numeric sensor takes time to execute
to add itself to shared memory. It causes that count of property in MRD
is smaller than expected. So, call updateReading() in ctor to add the
numeric sensor in shared memory even polling is not started yet.

tested:

1. run nv-shmem-tool read pldmd_HGX_CpuProcessorMetrics_0
2. check if all non-priority numeric sensor are listed

fixes nvbug https://nvbugs/5068397

Signed-off-by: Gilbert Chen <gilbertc@nvidia.com>

Co-authored-by: Aditya Kurdunkar <akurdunkar@nvidia.com>