Optimize OtherDeviceUpdater flow to improve performance
The ValidateDescriptor function in other_device_manager.cpp is being
called approximately 400 times. Within this function, we are making
synchronous calls to getDbusProperty, each of which takes about
2-5 ms for a round trip.

With this fix we are removing the redundant dbus calls.
The code processing time reduced to 125ms from 600ms.

Test performed -
1. Full bundle update
2. Target Filtering with 1 Retimer Update
3. Firmware Recovery
4. Debug Token Install/Erase
5. Update Retimer image which has only SKU as descriptor

Fixes JIRA https://

Signed-off-by: Rajeev Ranjan rajeeranjan@nvidia.com