Fix integrity check issue with corrupted signature header version
The integrity check currently allows modification of the minor and
major version fields in the signature header, making it possible to
install a firmware package with an incorrect version.

This fix immediately interrupts the update process if the minor or
major version is modified, preventing installation of a corrupted
firmware package.

Unit Tests:
- Modify the firmware package by changing the minor and major
  version to an invalid value (e.g., 0x04).
- Run a firmware update on a system with integrity check enabled.
- The update should fail with the correct Redfish error message.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5112823

Signed-off-by: Robert Papros <rpapros@nvidia.com>