mctp_discovery: Ignore MCTP endpoints that don't support PLDM
populateMctpInfo function does not return completion code,
so even if endpoint does not support PLDM it is registered
for D-Bus signals when endpoint is enabled or disabled. So
when these endpoints status changes GetPLDMTypes command is
send and failure is reported.

With the fix populateMctpInfo function returns true if
endpoint supports PLDM and is only registered for D-Bus
signals only in that case.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5074037

Signed-off-by: Tom Joseph <tjoseph@nvidia.com>