fw-update: Not fail fw update in PLDM UA when req data is corrupted
In the PLDM T5 flow, PLDM FD is the requester in the transfer, verify
and apply stage. In the HMC implementation, if the request data is
corrupted the state of the update is marked invalid by PLDM UA and
further requests from the PLDM FD is responded with 0x88
(COMMAND_NOT_EXPECTED) and eventually the update fails.

In the two cases the bug is reported, the corruption is seen with a
single RequestFwData and the subsequent retries from the PLDM FD is
valid request. So this patch adds a work around in PLDM UA, such that
corrupt request data in RequestFirmwareData does not cause FW update
to marked failed but retries are accepted, thus mitigating one-off
request data corruption.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5168572

Signed-off-by: Tom Joseph <tjoseph@nvidia.com>