Add support for InfiniBand PortProtocol
Entity type for InfiniBand and NVLink are same and NVLink is
proprietary.
Set the PortProtocol to "InfiniBand" if the EntityType is
PLDM_ENTITY_INFINIBAND in the PDR, only when the config is not
available in EM.

Tested without "PortProtocol" config for CX8 port and able to see
Ethernet and InfiniBand on redfish.
CX7 ManagementLinks and NVSwitch NVLinks show NVLink for PortProtocol
property.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5274674

Signed-off-by: Karthick Sundarrajan <ksundarrajan@nvidia.com>

Co-authored-by: Aditya Kurdunkar <akurdunkar@nvidia.com>