platform-mc: Get all properties at once from EM
pldm!566 introduced new properties Bus, Address in the SensorAuxName
interface in the EM configs for GB200. This was done so that two
identical entities on an IO Board can be recognized based on the i2c bus
addresses. Each of these properties (bus,address) are retrieved one by
one. This meant for config objects that don't have these properties an
error was logged; which caused a flood of logs.

This commit fixes this by retrieving all the properties of an object at
once and then checking for existence of these special properties.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5401943

Change-Id: Ibbbe929c80ad160461fffc48836c16dd743c8a7a
Signed-off-by: Aditya Kurdunkar <akurdunkar@nvidia.com>