platform-mc: Simplify device inventory creation
The existing flow for inventory creation uses the
`xyz.openbmc_project.Decorator.I2CDevice` interface to get the bus and
address for the terminus which it later uses to co-relate sensor
aux names for the terminus. Code Construct's mctpd no longer exposes the
decorator interface for the MCTP Endpoints.

The suggested alternative is to use EID directly for co-relation [1].
This patch simplifies the inventory creation based on the new changes.

[1]: https://gitlab-master.nvidia.com/dgx/bmc/openbmc/-/merge_requests/32892

Fixes nvbug https://nvbugspro.nvidia.com/bug/5426904

Change-Id: I86f4563d061d50d0381b6e0adca63c9ae8d32964
Signed-off-by: Aditya Kurdunkar <akurdunkar@nvidia.com>