fw-update: Move UA_T6 timer creation to end of RequestFWData
The current implementation does not have a timeout when waiting for
TransferComplete after the end of RequestFWData. This commit moves the
creation of timer from TransferComplete to when the last request is
made.

The timer also logs all the timeout failures as ApplyComplete timeouts,
this has been fixed to use the current PLDM state to create the correct
log.

Tests (using MockResponder):
- TransferComplete timeout is logged correctly
- ApplyComplete timeout is logged correctly
- VerifyComplete timeout is logged correctly

Fixes Jira https://

Signed-off-by: P Arun Kumar Reddy <parunkumarr@nvidia.com>